Adjusted navigation order back to original design. In order of how far you are navigating, and buttons size reduction. Perhaps we can get an option in this widget for ordering the buttons. Could be useful for switching up the order of non-navigation buttons in this widget as well.
Also I think we need to consider the user of Next and Previous in Chapter, we are treating and styling it like its similar to the chapter navigation button, but what it really is is a replacement for the Next and Previous buttons. As far as real world use of it, it acts exactly like the next and previous buttons for the chapter you are in, the only time it varies from the next previous is that it is disabled at the first and last comic of that chapter, therefor you would probably use it with chapter buttons, rather than with repetitive prev next buttons. The difference is if you have multiple series and then you would give them the option of next in series or next comic, again, both being next, and should probably duplicate the Next and Previous Graphic buttons or similar for the button images.