thumbnail change - comic post thumbnails will not display above the post for the comic post, the thumbnail is now used as an alternative for the latest thumbnail widget, if has_thumbnail the latest thumbnail widget will display that thumbnail inside the widget.
the_widget('ComicPressLatestThumbnailWidget','title=Latest Comic&onlyhome=true');

New option in the thumbnail 'onlyhome' will allow someone to have the thumbnail only display in on the homepage if widget enabled, off it will display everywhere it's placed.

Since it's a widget it can be used in site design for placement of an alternate image to the comic i.e. lfgcomic

Signed-off-by: Philip M. Hofer (Frumph) <frumph_dragon@yahoo.com>